Very basic implementation of API-compatible GNSS receiver added, documentation improved

+++ navigationSystem/src/quartzAccelerometer.cpp
@@ -2,13 +2,13 @@
 // Created by Tobias on 11/1/2024.
 //
 
-#include "accelerometer.hpp"
+#include "quartzAccelerometer.hpp"
 
 QuartzAccelerometer::QuartzAccelerometer() {
     measurementError = 0.1; // arbitrary value pending further research
 }
 
 // calculate single-axis linear acceleration
-float QuartzAccelerometer::getAcceleration() {
+double QuartzAccelerometer::getAcceleration() {
     return previousVelocity = (currentVelocity - previousVelocity)/interval;
 }
